Split out API call implementation to aid testing
Currently, the test for `Suspenseful` _must_ be given knowledge that it
is to be wrapped in `<React.Suspense>`. This creates a race condition in
the test when asserting the fallback value. Although this assertion
perhaps tests React, rather than the implementation here.

As an attempt to separate these concerns, the Promise resolution is
extracted & tested in isolation. The React component then only depends
on a function that produces data.