Implement composed components
The "standard" way of deciding to show a child component seems to be to
use something like "boolean driven development" & keeping the logic of
child components to render in the parent. This increases the parent's
responsibility & increases the required test coverage.

For example:

```ts
return (
  <Parent>
    {shouldShowChild && <Child/>}
    {!shouldShowChild && <OtherComponent/>}
  </Parent>
)
```

This commit inverts the control & injects the Child into the Parent
function. Testing is far more simple as each component has fewer
responsibilities.